+++ src/fr/maxlego08/menu/api/requirement/permissible/VaultPermissible.java
@@ -1,0 +1,9 @@
+package fr.maxlego08.menu.api.requirement.permissible;
+
+import fr.maxlego08.menu.api.requirement.Permissible;
+
+public interface VaultPermissible extends Permissible {
+
+    double getAmount();
+
+}