+++ changelog.md
@@ -42,6 +42,7 @@
 - Create action ``actionbar``, allows to send a message in the action bar of the player
 - Add placeholder ``%zmenu_player_is_expired_<key>%`` Check if a key has expired, if the key does not exist it will be considered expired
 - Add ``updateMasterButton``, allows to update the button entirely during the update. You have `update` enabled
+- Add requirement permissible ``money``, allows you to check if the player has enough money by using the Vault API
 - Fix folia (omg folia wtf)
 - Fix refresh action
 - Fix placeholder API cache, its disable by default